% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{all_param_codenames}
\alias{all_param_codenames}
\title{All Parameter and Method Codes in AQS data}
\format{A data frame with 1348 rows and 5 columns}
\description{
A dataset containing the unique combinations of Parameter and Method Codes
contained the in AQS observation data.
}

