% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{aqs06}
\alias{aqs06}
\title{AQS Measurements for 2006}
\format{A data frame with 188 rows and 9 columns
\describe{
  \item{SiteID}{Site ID of the monitoring station. The first }
  \item{Date}{Longitude of the site}
  \item{Species}{Latitude of the site}
  \item{Conc_obs}{horizontal coordinate of the site in Lambert Conformal Conical (LCC) projetion, in meters}
  \item{sig_c_obs}{vertical coordinate of the site in LCC projetion, in meters}
}}
\description{
A dataset containing the concentrations of forty-one chemical species,
as well as the uncertainties of those measurements, over the entire spatial
domain of the study.
}

