% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aqs_excel.R
\name{aqs_merge}
\alias{aqs_merge}
\title{Read in AQS data from folder of xlsx files}
\usage{
aqs_merge(fol_path, yr)
}
\arguments{
\item{fol_path}{file path (relative or complete) of the folder containing
the AQS files}

\item{year}{character vector of year the data corresponds to}
}
\value{
a dataframe of the aggregated data
}
\description{
Given a directory of AQS excel workbooks, reads the data from all the
sheets (named "UNC"), which contain uncertainty values, and the data from the
sheets (named "CONC") containing observed concentrations, and combines the
values all into a single dataframe.
}
\details{
Many assumptions are made concerning the naming of the files. For more
details, see the vignette describing the AQS data files
}

