% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{csn_site_index2}
\alias{csn_site_index2}
\title{Geographic data on 188 CSN sites}
\format{A data frame with 188 rows and 9 columns
\describe{
  \item{SiteID}{Site ID of the monitoring station. The first }
  \item{LON}{Longitude of the site}
  \item{LAT}{Latitude of the site}
  \item{Xm}{horizontal coordinate of the site in Lambert Conformal Conical (LCC) projetion, in meters}
  \item{Ym}{vertical coordinate of the site in LCC projetion, in meters}
  \item{StateCountySite}{the state code, county code, and site number of the
  monitoring site, separated by hyphens. This makes it easier to merge wtih AQS
  data available online.}
  \item{i,j}{respectively, the horizontal and vertical grid cell coordinates of the
  monitoring sites}
  \item{Network}{monitoring network the site is in}
}}
\source{
\url{http://www.epa.gov/ttn/airs/airsaqs/manuals/codedescs.htm}
}
\description{
A dataset containing information on the geographic location of Chemical
Speciation Network (CSN) sites. More detailed descriptions of the
fields/codes contained in it can be found via the url below (see "source").
}

