% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hybridSA.R
\name{get_optim}
\alias{get_optim}
\title{Get optimization values given asens and}
\usage{
get_optim(conc_by_day, sens_by_day, .sources)
}
\arguments{
\item{conc_by_day}{concentrations of chemical species for a given day}

\item{sens_by_day}{matrix of sensitivities, for a given day}

\item{.sources}{sources of PM2.5}
}
\value{
vector of length 20, either of NA's if input has any missing values,
 or numeric vector otherwise
}
\description{
Get optimization values given asens and
}
\details{
this is a function in beta
}

