% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aqs_excel.R
\name{getxl_sheet}
\alias{getxl_sheet}
\title{Extract data from an excel worksheet}
\usage{
getxl_sheet(wb_path, sheetname)
}
\arguments{
\item{wb_path}{character vector, path of the xl workbook to load data from}

\item{sheetname}{character vector, name of the sheet in the workbook to load
 data from}
}
\description{
Extract data from an excel worksheet
}

