% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hybridSA.R
\name{hybridsa}
\alias{hybridsa}
\title{Evaluate Hybrid SA Objective function}
\usage{
hybridsa(Rj, c_obs, csim, asens, sig_c_obs, sig_ctm, sig_lnr)
}
\arguments{
\item{Rj}{vector of R_j values. Has length equal to the number of sources.}

\item{asens}{2d matrix of average sensitivities, with rows corresponding to
species and columns corresponding to sources. The sensitivies are in the
CMAQ output, along with simulated concentrations}

\item{error}{difference in observations and simulated concentrations, given
by c_obs - c_init}

\item{w2}{vector of values eqtn{sigma^{-2}_{ln(R_j)}}.}

\item{weight}{the reciprocal of eqtn{sigma^2_{c_i^obs} + sigma^2_{SR_i^CTM}}; a
constant in the first summation term. sigma^2_{c_i^obs} is the uncertainty
in the observed concentration of species "i". eqtn{sigma^2_{SR_i^CTM}} is ...}

\item{gama}{value of the Gamma weight for the second summation in the Hybrid
 equation}
}
\description{
Describe this.
}
\details{
Gamma is a parameter specified by the user, and not necessarily
   bound to any observed or simulated data. The simulated concentrations
   are needed to obtain the error, which is given by c_obs - c_sim, the
   difference between the observed concentrations and the simulated
   concentrations.
}
\references{
\url{http://atmos-chem-phys.net/14/5415/2014/}
}

