% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{sig_ctm06}
\alias{sig_ctm06}
\title{Uncertainties for the optimization}
\source{
\url{http://www.diamondse.info/}
}
\description{
A named vector of length 41, with the \eqn{\sigma_{SR_i^CTM}} values for
each chemical species for 2006. These are taken to be constant throughout
the year. Note that these values need to be squared before being used in
the hybrid equation.
}
\references{
\url{http://atmos-chem-phys.net/14/5415/2014/}
}

