% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{sig_lnr06}
\alias{sig_lnr06}
\title{Uncertainties of ln(Rj) values}
\source{
\url{http://www.diamondse.info/}
}
\description{
An (un)named vector of length 20, with the \eqn{\sigma_{ln R_i}} values for
each source for 2006. These are taken to be constant throughout
the year. Note that these values need to be squared before being used in
the hybrid equation.
}
\references{
\url{http://atmos-chem-phys.net/14/5415/2014/}
}

