% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{used_paramcodes}
\alias{used_paramcodes}
\title{Parameter Codes Used in Selecting Observations}
\format{A data frame with 46 rows and 2 columns}
\description{
A dataset containing the parameter codes of the chemical species for which
the observational data was obtained.
}

