% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hybridSA.R
\name{hybridsa}
\alias{hybridsa}
\title{Evaluate Hybrid SA Objective function}
\usage{
hybridsa(Rj, c_obs, csim, asens, sig_c_obs, sig_ctm, sig_lnr)
}
\arguments{
\item{Rj}{vector of R_j values. Has length equal to the number of sources.}

\item{c_obs}{observed concentration}

\item{csim}{simulated concentration}

\item{asens}{2d matrix of average (daily) sensitivities, with rows corresponding to
species and columns corresponding to sources. The sensitivies are in the
CMAQ output, along with simulated concentrations}

\item{sig_c_obs}{uncertainties in observed concentrations}

\item{sig_ctm}{uncertainty in ....}

\item{sig_lnr}{uncertainty in Rj value (???)}
}
\description{
Describe this.
}
\details{
Gamma is a parameter specified by the user, and not necessarily
   bound to any observed or simulated data. The simulated concentrations
   are needed to obtain the error, which is given by c_obs - c_sim, the
   difference between the observed concentrations and the simulated
   concentrations.
}
\references{
\url{http://atmos-chem-phys.net/14/5415/2014/}
}

